Added package requirements.

+++ WordToVec/NeuralNetwork.py
@@ -13,7 +13,6 @@ import math
 
 
 class NeuralNetwork:
-
     __wordVectors: Matrix
     __wordVectorUpdate: Matrix
     __vocabulary: Vocabulary
@@ -36,6 +35,7 @@ class NeuralNetwork:
     parameter : WordToVecParameter
         Parameters of the Word2Vec algorithm.
     """
+
     def __init__(self, corpus: Corpus, parameter: WordToVecParameter):
         self.__vocabulary = Vocabulary(corpus)
         self.__parameter = parameter
@@ -48,6 +48,7 @@ class NeuralNetwork:
     Constructs the fast exponentiation table. Instead of taking exponent at each time, the algorithm will lookup
     the table.
     """
+
     def prepareExpTable(self):
         self.__expTable = [0.0] * (NeuralNetwork.EXP_TABLE_SIZE + 1)
         for i in range(NeuralNetwork.EXP_TABLE_SIZE):
@@ -63,6 +64,7 @@ class NeuralNetwork:
     VectorizedDictionary
         Dictionary of word vectors.
     """
+
     def train(self) -> VectorizedDictionary:
         result = VectorizedDictionary()
         if self.__parameter.isCbow():
@@ -90,17 +92,20 @@ class NeuralNetwork:
     float
         Calculated G value.
     """
+
     def calculateG(self, f: float, alpha: float, label: float) -> float:
         if f > NeuralNetwork.MAX_EXP:
             return (label - 1) * alpha
         elif f < -NeuralNetwork.MAX_EXP:
             return label * alpha
         else:
-            return (label - self.__expTable[((f + NeuralNetwork.MAX_EXP) * (NeuralNetwork.EXP_TABLE_SIZE // NeuralNetwork.MAX_EXP // 2))]) * alpha
+            return (label - self.__expTable[int((f + NeuralNetwork.MAX_EXP) *
+                                                (NeuralNetwork.EXP_TABLE_SIZE // NeuralNetwork.MAX_EXP // 2))]) * alpha
 
     """
     Main method for training the CBow version of Word2Vec algorithm.
     """
+
     def trainCbow(self):
         iteration = Iteration(self.__corpus, self.__parameter)
         currentSentence = self.__corpus.getSentence(iteration.getSentenceIndex())
@@ -132,7 +137,8 @@ class NeuralNetwork:
                         if f <= -NeuralNetwork.MAX_EXP or f >= NeuralNetwork.MAX_EXP:
                             continue
                         else:
-                            f = self.__expTable[((f + NeuralNetwork.MAX_EXP) * (NeuralNetwork.EXP_TABLE_SIZE // NeuralNetwork.MAX_EXP // 2))]
+                            f = self.__expTable[int((f + NeuralNetwork.MAX_EXP) *
+                                                    (NeuralNetwork.EXP_TABLE_SIZE // NeuralNetwork.MAX_EXP // 2))]
                         g = (1 - currentWord.getCode(d) - f) * iteration.getAlpha()
                         outputUpdate.addVector(self.__wordVectorUpdate.getRowVector(l2).product(g))
                         self.__wordVectorUpdate.addRowVector(l2, outputs.product(g))
@@ -163,6 +169,7 @@ class NeuralNetwork:
     """
     Main method for training the SkipGram version of Word2Vec algorithm.
     """
+
     def trainSkipGram(self):
         iteration = Iteration(self.__corpus, self.__parameter)
         currentSentence = self.__corpus.getSentence(iteration.getSentenceIndex())
@@ -191,7 +198,8 @@ class NeuralNetwork:
                             if f <= -NeuralNetwork.MAX_EXP or f >= NeuralNetwork.MAX_EXP:
                                 continue
                             else:
-                                f = self.__expTable[((f + NeuralNetwork.MAX_EXP) * (NeuralNetwork.EXP_TABLE_SIZE // NeuralNetwork.MAX_EXP // 2))]
+                                f = self.__expTable[int((f + NeuralNetwork.MAX_EXP) *
+                                                        (NeuralNetwork.EXP_TABLE_SIZE // NeuralNetwork.MAX_EXP // 2))]
                             g = (1 - currentWord.getCode(d) - f) * iteration.getAlpha()
                             outputUpdate.addVector(self.__wordVectorUpdate.getRowVector(l2).product(g))
                             self.__wordVectorUpdate.addRowVector(l2, self.__wordVectors.getRowVector(l1).product(g))
